+++ nav_bringup/config/localization.yaml
@@ -1,0 +1,88 @@
+ekf_local:
+  ros__parameters:
+    frequency: 50.0
+    sensor_timeout: 0.1
+    two_d_mode: true
+    publish_tf: true
+    print_diagnostics: true
+    debug: true
+
+    twist0: enc_vel/raw
+    twist0_config: [false, false, false,
+                    false, false, false,
+                    true,  false, false,
+                    false, false, true,
+                    false, false, false]
+    twist0_queue_size: 50
+    twist0_nodelay: true
+    twist0_differential: false
+    twist0_relative: false
+
+    imu0: imu/raw
+    imu0_config: [false, false, false,
+                  false, false, false,
+                  false, false, false,
+                  false, false, true,
+                  false, false, false]
+    imu0_queue_size: 10
+    imu0_nodelay: false
+    imu0_differential: false
+    imu0_relative: false
+    imu0_remove_gravitational_acceleration: true
+
+
+navsat_transform:
+  ros__parameters:
+    frequency: 20.0
+    delay: 0.0
+    magnetic_declination_radians: -0.1269
+    yaw_offset: 0.0
+    zero_altitude: true
+    broadcast_utm_transform: false
+    publish_filtered_gps: false
+    use_odometry_yaw: false
+    wait_for_datum: true
+
+
+ekf_global:
+  ros__parameters:
+    frequency: 30.0
+    sensor_timeout: 1.5
+    two_d_mode: true
+    publish_tf: true
+    print_diagnostics: true
+    debug: true
+
+    odom0: odom/local
+    odom0_config: [true,  true,  false,
+                   false, false, true,
+                   true,  false, false,
+                   false, false, true,
+                   false, false, false]
+    odom0_queue_size: 10
+    odom0_nodelay: true
+    odom0_differential: false
+    odom0_relative: false
+
+    odom1: odom/gps
+    odom1_config: [true,  true,  false,
+                   false, false, false,
+                   false, false, false,
+                   false, false, false,
+                   false, false, false]
+    odom1_queue_size: 10
+    odom1_nodelay: true
+    odom1_differential: false
+    odom1_relative: false
+
+    imu0: imu/raw
+    imu0_config: [false, false, false,
+                  false, false, true,
+                  false, false, false,
+                  false,  false, false,
+                  false,  false, false]
+    imu0_queue_size: 10    
+    imu0_nodelay: true
+    imu0_differential: false
+    imu0_relative: false
+    imu0_remove_gravitational_acceleration: true
